+++ BusinessLogic/Domain/ProductItem.cs
@@ -5,14 +5,14 @@ namespace BusinessLogic.Domain
 {
     public class ProductItem
     {
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public int? Weight { get; set; }
         public decimal UnitPrice { get; set; }
         public decimal Discount { get; set; }
         public Product Product { get; set; }
         public AuditInfo? AuditInfo { get; set ; }
 
-        public ProductItem(int quantity, int? weight, decimal unitPrice, decimal discount, Product product)
+        public ProductItem(decimal quantity, int? weight, decimal unitPrice, decimal discount, Product product)
         {
             Quantity = quantity;
             Weight = weight;
